requirements.txt
streamlit
google-generativeai
reportlab
python-dotenv
streamlit-mic-recorder
